User authentication & authorization

+++ src/main/java/com/apiit/eeashopping/Controllers/CategoryController.java
@@ -3,7 +3,10 @@ package com.apiit.eeashopping.Controllers;
 import com.apiit.eeashopping.DB.CategoryRepository;
 import com.apiit.eeashopping.Model.Category;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 @RestController
 @RequestMapping(value = "/category")
@@ -12,22 +15,25 @@ public class CategoryController {
     @Autowired
     CategoryRepository categoryRepository;
 
-    @GetMapping(path = "/all")
+    @GetMapping(path = "/public/all")
     public @ResponseBody Iterable<Category> getAllCategories(){
         System.out.println("Fetching all categories");
         return categoryRepository.findAll();
     }
 
-    @PostMapping(path = "/add")
-    public Category addNewCategory(@RequestBody Category category){
-
+    @PreAuthorize("hasAnyRole('ADMIN')")
+    @PostMapping(path = "/admin/add")
+    public Category addNewCategory(@RequestBody String categoryName){
+        Category category = new Category();
+        category.setName(categoryName);
         categoryRepository.save(category);
-        System.out.println(category.getName() + "category is added");
+        System.out.println(category.getName() + " category is added");
 
         return category;
     }
 
-    @PutMapping
+    @PreAuthorize("hasAnyRole('ADMIN')")
+    @PutMapping(path = "/admin")
     public Category updateCategory(@RequestBody Category category){
 
         categoryRepository.save(category);
@@ -36,13 +42,15 @@ public class CategoryController {
         return category;
     }
 
-    @DeleteMapping("/{catid}")
+    @PreAuthorize("hasAnyRole('ADMIN')")
+    @DeleteMapping("/admin/{catid}")
     public boolean deleteCategory(@PathVariable int catId){
 
-        Category c = categoryRepository.findById(catId).get();
+        Optional<Category> c = categoryRepository.findById(catId);
+        if(c.isPresent())
         categoryRepository.deleteById(catId);
-        System.out.println(c.getName() + " is deleted");
+        System.out.println(c.get().getName() + " is deleted");
 
-        return true;
+        return c.isPresent();
     }
 }
